services/bot/requirements.txt
aiogram==2.25.1
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiodns==3.1.1
pycares==4.4.0
requests==2.31.0
urllib3==1.26.18
sqlalchemy==2.0.26
pillow==9.5.0 